Fix spacing and margins
Margin is outside of the box, so putting it on the grid columns was
pushing the total row to be too big in large mode, causing the row to
break.  Needed to put the margin (and background color) in the content p
inside the grid.  Also fix the width of the third box in tablet mode to
fill the screen.